# Migrate to new ArrayBuffer