--- a/Package.resolved
+++ b/Package.resolved
@@ -6,7 +6,7 @@
         "repositoryURL": "https://github.com/saeta/penguin.git",
         "state": {
           "branch": "master",
-          "revision": "1586b9b88628fb772e7b1d6e79185dcf5e3d2c6d",
+          "revision": "4ea37a85ab75ddd5f02c395b9da535d6cbd66f20",
           "version": null
         }
       },
@@ -24,7 +24,7 @@
         "repositoryURL": "https://github.com/google/swift-benchmark.git",
         "state": {
           "branch": "master",
-          "revision": "66262f4f85cc2cd319f4aaae33368a14ea1c1dd9",
+          "revision": "3d24a938ea553a3977cfbf8199d42839c9061c3f",
           "version": null
         }
       },