… Sources/SwiftFusion/Inference/FactorsStorage.swift
… Sources/SwiftFusion/Inference/ValuesStorage.swift
… Tests/SwiftFusionTests/Inference/VariableAssignmentsTests.swift