Migrates everything to the new `ArrayBuffer` interface from https://github.com/saeta/penguin/pull/82.

The boilerplate in `ValuesStorage` and `FactorsStorage` feels simpler now, which is good!

Benchmarks before:
```
swift build -c release -Xswiftc -cross-module-optimization && .build/release/SwiftFusionBenchmarks --filter NewFactorGraph
name                                                                   time           std        iterations
-----------------------------------------------------------------------------------------------------------
Pose2SLAM.NewFactorGraph, Intel, 10 Gauss-Newton steps, 500 CGLS steps  456.792364 ms ±   0.00 %          1
Pose3SLAM.NewFactorGraph, sphere2500, 30 LM steps, 200 CGLS steps      5922.074432 ms ±   0.00 %          1
```

Benchmarks after:
```
swift build -c release -Xswiftc -cross-module-optimization && .build/release/SwiftFusionBenchmarks --filter NewFactorGraph
name                                                                   time           std        iterations
-----------------------------------------------------------------------------------------------------------
Pose2SLAM.NewFactorGraph, Intel, 10 Gauss-Newton steps, 500 CGLS steps  446.596137 ms ±   0.00 %          1
Pose3SLAM.NewFactorGraph, sphere2500, 30 LM steps, 200 CGLS steps      4950.754052 ms ±   0.00 %          1
```